horizontal_fl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "horizontal-fl-chemsafe"
version = "0.1.0"
description = "Horizontal Federated Learning for Chemical Safety Prediction"
dependencies = [
    "flower>=1.23.0",
    "pandas>=2.2.3",
    "numpy>=1.26.4,<2.0",
    "scikit-learn>=1.6.1",
    "xgboost>=3.0.2",
    "rdkit>=2025.3.2",
    "imbalanced-learn>=0.13.0",
    "molvs>=0.1.1",
    "plotly>=6.1.2",
    "umap-learn>=0.5.7",
    "kaleido>=0.2.1",
    "matplotlib>=3.10.7",
    "statsmodels>=0.14.5",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "chemfl"

[tool.flwr.app.components]
serverapp = "server_app:app"
clientapp = "client_app:app"

[tool.flwr.app.config]
num-server-rounds = 10
epochs = 5

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 3
